backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
mediapipe==0.10.21
opencv-python-headless==4.10.0.84
numpy
pydantic==2.10.4
python-multipart
Pillow
httpx>=0.27.0
python-dotenv>=0.9.9
